Renamed make_enum_macro to just make_enum
