+++ src/software/ai/motion_constraint/motion_constraint.h
@@ -2,7 +2,7 @@
 
 #include <ostream>
 
-#include "software/util/make_enum_macro/make_enum_macro.h"
+#include "software/util/make_enum/make_enum.h"
 
 /**
  * This enum describes the constraints robot movement. They are
