+++ src/software/sensor_fusion/BUILD
@@ -33,7 +33,7 @@ cc_library(
         "//software/geom",
         "//software/proto:ssl_cc_proto",
         "//software/time:timestamp",
-        "//software/util/make_enum_macro",
+        "//software/util/make_enum",
     ],
 )
