+++ src/software/util/make_enum/make_enum_test.cpp
@@ -1,36 +1,36 @@
-#include "software/util/make_enum_macro/make_enum_macro.h"
+#include "software/util/make_enum/make_enum.h"
 
 #include <gtest/gtest.h>
 
-TEST(PrintableEnumMacroTest, validate_valid_enum_args_without_trailing_comma)
+TEST(EnumMacroTest, validate_valid_enum_args_without_trailing_comma)
 {
     std::string enum_string = "foo, bar, baz";
     bool result             = isEnumArgsValid(enum_string);
     EXPECT_TRUE(result);
 }
 
-TEST(PrintableEnumMacroTest, validate_valid_enum_args_with_trailing_comma)
+TEST(EnumMacroTest, validate_valid_enum_args_with_trailing_comma)
 {
     std::string enum_string = "foo, bar, baz,";
     bool result             = isEnumArgsValid(enum_string);
     EXPECT_TRUE(result);
 }
 
-TEST(PrintableEnumMacroTest, validate_invalid_enum_args)
+TEST(EnumMacroTest, validate_invalid_enum_args)
 {
     std::string enum_string = "foo, bar=3, baz,";
     bool result             = isEnumArgsValid(enum_string);
     EXPECT_FALSE(result);
 }
 
-TEST(PrintableEnumMacroTest, separate_enum_strings_empty_string)
+TEST(EnumMacroTest, separate_enum_strings_empty_string)
 {
     std::string enum_string = "";
     auto result             = separateEnumStrings(enum_string);
     EXPECT_TRUE(result.empty());
 }
 
-TEST(PrintableEnumMacroTest, separate_enum_strings_without_trailing_comma)
+TEST(EnumMacroTest, separate_enum_strings_without_trailing_comma)
 {
     std::string enum_string                  = "foo, bar, baz";
     auto result                              = separateEnumStrings(enum_string);
@@ -38,7 +38,7 @@ TEST(PrintableEnumMacroTest, separate_enum_strings_without_trailing_comma)
     EXPECT_EQ(expected_result, result);
 }
 
-TEST(PrintableEnumMacroTest, separate_enum_strings_with_extra_spaces_tabs_and_newlines)
+TEST(EnumMacroTest, separate_enum_strings_with_extra_spaces_tabs_and_newlines)
 {
     std::string enum_string                  = "foo   \n, \tbar, \nbaz \t ";
     auto result                              = separateEnumStrings(enum_string);
@@ -46,7 +46,7 @@ TEST(PrintableEnumMacroTest, separate_enum_strings_with_extra_spaces_tabs_and_ne
     EXPECT_EQ(expected_result, result);
 }
 
-TEST(PrintableEnumMacroTest, separate_enum_strings_with_trailing_comma)
+TEST(EnumMacroTest, separate_enum_strings_with_trailing_comma)
 {
     std::string enum_string                  = "foo, bar, baz,";
     auto result                              = separateEnumStrings(enum_string);
@@ -55,7 +55,7 @@ TEST(PrintableEnumMacroTest, separate_enum_strings_with_trailing_comma)
 }
 
 MAKE_ENUM(TestEnum, Var1, Var2)
-TEST(PrintableEnumMacroTest, test_macro_generated_functions)
+TEST(EnumMacroTest, test_macro_generated_functions)
 {
     EXPECT_EQ(2, sizeTestEnum());
     EXPECT_EQ(std::vector<TestEnum>({TestEnum::Var1, TestEnum::Var2}),
